revert refactor to tooltips
