+++ src/main/java/net/dries007/tfc/client/ClientHelpers.java
@@ -6,10 +6,8 @@
 
 package net.dries007.tfc.client;
 
-import java.util.List;
 import javax.annotation.Nullable;
 
-import net.minecraft.ChatFormatting;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.client.model.geom.ModelLayerLocation;
@@ -20,20 +18,13 @@ import net.minecraft.client.renderer.block.model.ItemTransforms;
 import net.minecraft.client.renderer.entity.player.PlayerRenderer;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.Direction;
-import net.minecraft.network.chat.Component;
-import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.util.Mth;
 import net.minecraft.world.entity.HumanoidArm;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.item.TooltipFlag;
-import net.minecraft.world.level.BlockGetter;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.phys.BlockHitResult;
-
-import net.minecraftforge.fluids.FluidStack;
-import net.minecraftforge.items.IItemHandler;
 
 import com.mojang.blaze3d.systems.RenderSystem;
 import com.mojang.blaze3d.vertex.PoseStack;
@@ -133,46 +124,6 @@ public final class ClientHelpers
             right ? ItemTransforms.TransformType.FIRST_PERSON_RIGHT_HAND : ItemTransforms.TransformType.FIRST_PERSON_LEFT_HAND,
             !right, poseStack, source, combinedLight);
 
-    }
-
-    /**
-     * Adds a tooltip based on an inventory, listing out the items inside.
-     * Modified from {@link net.minecraft.world.level.block.ShulkerBoxBlock#appendHoverText(ItemStack, BlockGetter, List, TooltipFlag)}
-     */
-    public static void addInventoryTooltipInfo(IItemHandler inventory, List<Component> tooltips)
-    {
-        int maximumItems = 0, totalItems = 0;
-        for (ItemStack stack : Helpers.iterate(inventory))
-        {
-            if (!stack.isEmpty())
-            {
-                ++totalItems;
-                if (maximumItems <= 4)
-                {
-                    ++maximumItems;
-                    tooltips.add(stack.getHoverName().copy()
-                        .append(" x")
-                        .append(String.valueOf(stack.getCount())));
-                }
-            }
-        }
-
-        if (totalItems - maximumItems > 0)
-        {
-            tooltips.add(new TranslatableComponent("container.shulkerBox.more", totalItems - maximumItems).withStyle(ChatFormatting.ITALIC));
-        }
-    }
-
-    /**
-     * Adds a tooltip based on a single fluid stack
-     */
-    public static void addFluidStackTooltipInfo(FluidStack fluid, List<Component> tooltips)
-    {
-        if (!fluid.isEmpty())
-        {
-            tooltips.add(new TranslatableComponent("tfc.tooltip.fluid_units_of", fluid.getAmount())
-                .append(fluid.getDisplayName()));
-        }
     }
 
     private static float calculateTilt(float pitch)
